fixing bug

+++ app/Http/Controllers/BookController.php
@@ -109,7 +109,9 @@ class BookController extends Controller
         $id = Crypt::decrypt($id);
         $item = Book::find($id);
         if ($item) {
-            Storage::delete($item->dokumen);
+            if ($item->dokumen != null) {
+                Storage::delete($item->dokumen);
+            }            
             $item->delete();
             return response()->json(['success' => 'Item deleted successfully.']);
         } else {
